Restore stops. Create variations.
